+++ packages/components/test/index.js
@@ -341,6 +341,21 @@ describe('Select', () => {
     )
     expect(json).toMatchSnapshot()
   })
+
+  test('renders with custom icon', () => {
+    const json = renderJSON(
+      <ThemeProvider theme={theme}>
+        <Select
+          arrow={
+            <svg>
+              <path d="M7.41 7.84l4.59 4.58 4.59-4.58 1.41 1.41-6 6-6-6z" />
+            </svg>
+          }
+        />
+      </ThemeProvider>
+    )
+    expect(json).toMatchSnapshot()
+  })
 })
 
 describe('Textarea', () => {